% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNoise.R
\name{addNoise}
\alias{addNoise}
\title{Add Noise}
\usage{
addNoise(x, input.folder = "", output.folder = "", expIds = c(),
  add = FALSE, scaling.factor = 1, noise.type = "pois", param1 = 1,
  param2 = 1, scale.noise = 1)
}
\arguments{
\item{x}{matrix with taxa as rows and time points as columns (optional, can be set to NULL if an input.folder is provided)}

\item{input.folder}{folder from which to read time series (optional, input folder should have subfolders as generated by function generateTS.R)}

\item{output.folder}{folder to which modified time series are written (optional)}

\item{expIds}{set of experiment identifiers to process (only needed if input.folder is provided)}

\item{add}{whether to sample the abundance value from the noise distribution (the default) or whether to add noise, where the noise value is sampled from the noise distribution with user-specified parameters}

\item{scaling.factor}{if unequal to one: scale observed values with the given scaling factor and round (rounding will convert non-count data into counts)}

\item{noise.type}{which noise distribution to use (pois = Poisson, the default, nb = negative binomial, norm = Gaussian)}

\item{param1}{first parameter for noise distribution (when add is false: standard deviation for Gaussian distribution, success probability for negative binomial, when add is true: mean for Gaussian distribution, lambda for Poisson, size for negative binomial)}

\item{param2}{second parameter for noise distribution (only when add is true: standard deviation for Gaussian distribution, success probability for negative binomial)}

\item{scale.noise}{divide noise value by scale.noise before adding it to the observed value (only when add is true)}
}
\value{
if x is provided without an output.folder, the modified matrix is returned; if input.folder is provided without an output folder, the last processed experiment identifier is returned
}
\description{
Add noise to time series.
}
\details{
Noise is either generated by setting a noise parameter to the observed value and sampling a new value from this noise distribution (add=FALSE)
or by sampling the selected noise distribution with user-provided parameters and adding the noise value to each time point independently (add=TRUE).
When choosing to add Gaussian noise and the resulting value is negative, it will be set to zero.
When generating modified values with the negative binomial and add set to false, the resulting value is added to the observed value.
It is not recommended to enable add when using the Poisson or negative binomial distribution.
}
\examples{
  N <- 50
  metapop <- generateAbundances(N=500, mode=5, probabs=TRUE)
  ts <- simHubbell(N=N, M=500,I=1500,d=N, m.vector=metapop, tskip=500, tend=1000)
  ts.noisy <- addNoise(ts)
}
