% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustc.R
\name{adjustc}
\alias{adjustc}
\title{Adjust connectance}
\usage{
adjustc(A, c = 0.02)
}
\arguments{
\item{A}{interaction matrix}

\item{c}{target connectance}
}
\value{
adjusted interaction matrix
}
\description{
The connectance is adjusted by randomly adding or removing interactions.
}
\examples{
A=generateA(N=10)
A.adj=adjustc(A,c=0.5)
}

