% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignTaxonLevelsToA.R
\name{assignTaxonLevelsToA}
\alias{assignTaxonLevelsToA}
\title{Assign Taxon Levels}
\usage{
assignTaxonLevelsToA(A, type = "stoola", taxon.level = "genus",
  uniqueNames = FALSE, higherLevelNames = TRUE)
}
\arguments{
\item{A}{an interaction matrix inferred from the David data processed with method generateTS}

\item{type}{the data set, supported: stoola and stoolb}

\item{taxon.level}{the taxon level to be assigned as row and column names, supported: otu, species, genus, family, order, class, phylum}

\item{uniqueNames}{make names unique by appending a counter if needed}

\item{higherLevelNames}{if given level is not known, assign the highest level that is known}
}
\description{
Assign OTU identifiers or higher-level taxon names to an interaction matrix inferred from the David data processed with method generateTS.
}
\examples{
  # TODO: define stoolA in this example
  \dontrun{stoolA <- assignTaxonLevelsToA(stoolA, type = "stoola")}
}
