% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorVsTaxonNum.R
\name{autocorVsTaxonNum}
\alias{autocorVsTaxonNum}
\title{Compute slope for increasing number of taxa versus their autocorrelation}
\usage{
autocorVsTaxonNum(ts, lag = 1, plot = FALSE, col = "black",
  pch = "+")
}
\arguments{
\item{ts}{a matrix with objects as rows and time points as columns}

\item{lag}{the lag for which autocorrelation is computed}

\item{plot}{plot the increasing number of taxa versus their autocorrelation}

\item{col}{plot color}

\item{pch}{plot point character}
}
\value{
a list with the taxon number, autocorrelations, slope and p-value
}
\description{
The taxa are added in decreasing order of abundance.
}
\examples{
\dontrun{
N=10
A=generateA(N,c=0.1)
soi.ts=soi(N=N,I=1000,A=A,tend=100)
res=autocorVsTaxonNum(soi.ts, plot=TRUE)
}
}
