% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binAutocorrelations.R
\name{binAutocorrelations}
\alias{binAutocorrelations}
\title{Bin taxa by maximal autocorrelation}
\usage{
binAutocorrelations(x, thresholds = c(0.3, 0.5, 0.8))
}
\arguments{
\item{x}{the time series with taxa as rows and samples as columns}

\item{thresholds}{binning thresholds}
}
\value{
a list with the indices of taxa in each autocorrelation bin
}
\description{
The taxa are binned according to their maximal autocorrelation into
4 categories with low, middle, high and very high autocorrelation.
The autocorrelation at lag = 0 is ignored. The binning thresholds
can be altered.
}
\examples{
out=binAutocorrelations(simuntb(N=100))
length(out$lowmaxautocor)
length(out$highmaxautocor)
length(out$veryhighmaxautocor)
}

