% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binHurst.R
\name{binHurst}
\alias{binHurst}
\title{Given a community time series, bin taxa according to their Hurst exponent}
\usage{
binHurst(x, thresholds = c(0.5, 0.7, 0.9))
}
\arguments{
\item{the}{time series with taxa as rows and samples as columns}
}
\value{
the indices of the taxa in each Hurst category, labeled lowH, middleH, highH and veryhighH
}
\description{
Community taxa can be placed into a low, middle, high and very high Hurst category. The
thresholds between categories can be user-defined.
}
\examples{
out=binHurst(simuntb(N=100))
length(out$veryhighH)
length(out$highH)
length(out$middleH)
length(out$lowH)
}

