% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.noisetypes.R
\name{boxplot.noisetypes}
\alias{boxplot.noisetypes}
\title{Do a boxplot of matrix rows classified by noise type}
\usage{
\method{boxplot}{noisetypes}(x, noisetypes)
}
\arguments{
\item{x}{a matrix}

\item{noisetypes}{the noisetypes object generated for the matrix}
}
\description{
Do a boxplot of matrix rows classified by noise type
}

