% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caporaso_F4TongueL6.R
\docType{data}
\name{caporaso_F4TongueL6}
\alias{caporaso_F4TongueL6}
\title{Caporaso Tongue Sequencing Data for Subject F4 on level 6}
\format{An object of class \code{data.frame} with 373 rows and 135 columns.}
\usage{
data(caporaso_F4TongueL6)
}
\value{
TaxonxSample matrix of size 373 x 135
}
\description{
Mostly genus-level counts (L6) from tongue samples of subject F4 over time.
}
\details{
Subject F4 was followed over 6 months. V4 16S region was sequenced. A counter was attached to genera appearing more than once in the count table.
}
\examples{
data(caporaso_F4TongueL6)
# sort taxa by abundance and print top 10 most abundant taxa
sorted <- sort(apply(caporaso_F4TongueL6,1,sum),decreasing = TRUE, index.return = TRUE)
rownames(caporaso_F4TongueL6)[sorted$ix[1:10]]
}
\references{
Caporaso et al. (2011) Moving pictures of the human microbiome Genome Biology vol. 12:R50
\href{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2011-12-5-r50}{Genome Biology}
}
\keyword{data,}
\keyword{datasets}
