% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caporaso_M3FecesL6.R
\docType{data}
\name{caporaso_M3FecesL6}
\alias{caporaso_M3FecesL6}
\title{Caporaso Stool Sequencing Data for Subject M3 on level 6}
\format{An object of class \code{data.frame} with 335 rows and 332 columns.}
\usage{
data(caporaso_M3FecesL6)
}
\value{
TaxonxSample matrix of size 335 x 332
}
\description{
Mostly genus-level counts (L6) from stool samples of subject M3 over time.
}
\details{
Subject M3 was followed for 15 months. V4 16S region was sequenced. A counter was attached to genera appearing more than once in the count table.
}
\examples{
data(caporaso_M3FecesL6)
# sort taxa by abundance and print top 10 most abundant taxa
sorted <- sort(apply(caporaso_M3FecesL6,1,sum),decreasing = TRUE, index.return = TRUE)
rownames(caporaso_M3FecesL6)[sorted$ix[1:10]]
}
\references{
Caporaso et al. (2011) Moving pictures of the human microbiome Genome Biology vol. 12:R50
\href{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2011-12-5-r50}{Genome Biology}
}
\keyword{data,}
\keyword{datasets}
