% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFit.R
\name{compareFit}
\alias{compareFit}
\title{Collect fitting results to community time series.
Different quality scores are computed for each of the three
fitting methods: limits, neutral and soc.
If the input settings are provided via the input.folder argument,
the parameters obtained through fitting are compared to the
real parameters.}
\usage{
compareFit(fit.folder = "", input.folder = "", expIds = c(),
  fit.method = "limits", fit.type = "all")
}
\arguments{
\item{fit.folder}{the folder where fitting results are stored.}

\item{input.folder}{(optional) the folder where results of function generateTS are stored.}

\item{expIds}{the experiment identifiers of time series to be considered}

\item{fit.method}{the fitting method (limits, neutral or soc)}

\item{fit.type}{the part of the time series used (all, selected, first)}
}
\value{
a table with goodness of fit scores
}
\description{
Collect fitting results to community time series.
Different quality scores are computed for each of the three
fitting methods: limits, neutral and soc.
If the input settings are provided via the input.folder argument,
the parameters obtained through fitting are compared to the
real parameters.
}

