% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david_stoolA_metadata.R
\docType{data}
\name{david_stoolA_metadata}
\alias{david_stoolA_metadata}
\title{David et al. stool sample metadata object of subject A}
\format{a matrix with 12 rows representing metadata and 329 columns representing samples}
\usage{
data(david_stoolA_metadata)
}
\description{
The metadata collected on stool samples of subject A over time.
Subject A traveled from day 71 to day 122 and contracted diarrhea on
days 80 to 85 and 104 to 113.
}
\examples{
data(david_stoolA_metadata)
}
\references{
David et al. (2014) Host lifestyle affects human microbiota on daily timescales Genome Biology vol. 15 (7):R89
\href{http://www.genomebiology.com/2014/15/7/R89}{Genome Biology}
}
\keyword{datasets}
