% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david_stoolB_metadata.R
\docType{data}
\name{david_stoolB_metadata}
\alias{david_stoolB_metadata}
\title{David et al. stool sample metadata object of subject B}
\format{a matrix with 2 rows representing metadata and 190 columns representing samples}
\usage{
data(david_stoolB_metadata)
}
\description{
The metadata collected on stool samples of subject B over time.
In contrast to subject A, these include only collection day and age.
Salmonella infection (food poisoning) peaked at day 159.
}
\examples{
data(david_stoolB_metadata)
}
\references{
David et al. (2014) Host lifestyle affects human microbiota on daily timescales Genome Biology vol. 15 (7):R89
\href{http://www.genomebiology.com/2014/15/7/R89}{Genome Biology}
}
\keyword{datasets}
