% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david_stoolB_otus.R
\docType{data}
\name{david_stoolB_otus}
\alias{david_stoolB_otus}
\title{David Stool Sequencing Data for Subject B}
\format{An object of class \code{data.frame} with 5431 rows and 190 columns.}
\usage{
data(david_stoolB_otus)
}
\value{
#' @return OTUxSample matrix of size 5431 x 190
}
\description{
The OTUs sequenced from stool samples of subject B over time.
}
\details{
Salmonella infection (food poisoning) peaked at day 159.
}
\examples{
data(david_stoolB_otus)
# Sort OTUs by abundance
sorted <- sort(apply(david_stoolB_otus,1,sum),
                 decreasing=TRUE,index.return=TRUE)
# Display the top 10 most abundant OTUs
tsplot(david_stoolB_otus[sorted$ix[1:10],], type = "l", header = "Stool B")
}
\references{
David et al. (2014) Host lifestyle affects human microbiota on daily timescales Genome Biology vol. 15 (7):R89
\href{http://www.genomebiology.com/2014/15/7/R89}{Genome Biology}
}
\keyword{data,}
\keyword{datasets}
