% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envGrowthChanges.R
\name{envGrowthChanges}
\alias{envGrowthChanges}
\title{Generate growth changes induced by environment}
\usage{
envGrowthChanges(species, env.factors = 2, conditions = 2, strength)
}
\arguments{
\item{species}{Number of species}

\item{env.factors}{Number of environmental factors}

\item{conditions}{Number of environmental conditions}

\item{strength}{Strength of environmental factors}
}
\value{
Matrix of growth changes, 1 column per condition.
}
\description{
Generate a matrix of growth changes for
different environmental conditions. This vector can be supplied to generateDataSet to introduce environmental perturbation in the dataset.
Note that the effect of environmental parameters is already summed, so only one growth change is provided per species per condition.
}
\examples{
klemm = generateA(N=10, type="klemm", c=0.5)
env = envGrowthChanges(species = 10, env.factors=2, conditions=2, strength=0.5)
dataset = generateDataSet(100, klemm, env.matrix = env, perturb.count = c(50, 50))
}
