% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataSet.R
\name{generateDataSet}
\alias{generateDataSet}
\title{Generate a dataset}
\usage{
generateDataSet(samples, matrix, env.matrix = NULL, perturb.count = NULL,
  perturb = NULL, count = 1000, mode = 4)
}
\arguments{
\item{samples}{Number of samples}

\item{matrix}{Interaction matrix (generated from generateA.R)}

\item{env.matrix}{Growth rate changes induced by environment; 1 column per environmental condition.}

\item{perturb.count}{Number of samples per environmental condition. Sum should be equal to total number of samples.}

\item{perturb}{Perturbation object. At the moment, only changes in growth rates are supported.}

\item{count}{Total number of individuals in dataset}

\item{mode}{Mode for generateAbundances; default value samples counts from Poisson distribution with lambda count/N}
}
\value{
The abundance dataset
}
\description{
Generate an abundance dataset, with or without environmental perturbance, by using generalized Lotka-Volterra.
For each sample, a time series is generated with Lotka-Volterra and sampled at t=1000.
The environmental perturbances can be generated with envGrowthChanges
}
