% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTS.R
\name{generateTS}
\alias{generateTS}
\title{Generate community time series}
\usage{
generateTS(N = 100, I = 1500, tend = 100, initAbundMode = 5, c = 0.02,
  Atype = "klemm", d = -1, PEP = 30, m = 0.05, algorithm = "dm",
  interval = 1, output.folder = "", expId = "", maxIter = 10)
}
\arguments{
\item{N}{species number}

\item{I}{individual number (only for SOC and UNTB)}

\item{tend}{the number of time points to generate}

\item{initAbundMode}{the mode in which initial abundances are generated (described in function generateAbundances)}

\item{c}{connectivity of interaction matrix (only for SOC, ricker and gLV)}

\item{Atype}{method to generate the interaction matrix (described in function generateA)}

\item{d}{diagonal value of the interaction matrix}

\item{PEP}{positive edge percentage of interaction matrix (only for SOC, ricker and gLV)}

\item{m}{immigration rate (only for UNTB)}

\item{algorithm}{how time series should be generated (can be untb, soc, ricker, glv, dm, davida or davidb)}

\item{interval}{sampling frequency (if 1, each sample is taken, if 2, every second sample is taken etc.)}

\item{output.folder}{path to result folder, if specified, input parameter and time series are exported to the output folder}

\item{expId}{the identifier of the time series generation experiment}

\item{maxIter}{maximal iteration number for finding a stable interaction matrix}
}
\value{
list containing the generated interaction matrix, initial abundances (they double as immigration probabilities for soc), carrying capacities (they double as growth rates for glv), extinction probabilities, time series and settings
}
\description{
Wrapper function for time series generation.
The result folder has a specific structure that is
created if it does not yet exist. It has two
sub-folders, one named timeseries and the other
settings. In both sub-folders, there are folders
having as name the experiment identifier followed by
timeseries or settings. The interaction matrix,
immigration and extinction rates, carrying capacities
and other randomly generated input are also
stored in settings. The idea is that the time
series can be re-created from the information
stored in the settings folder (though not
exactly, since the SOC, Ricker and neutral models have random components).
Furthermore, a plot of the time series is
stored in settings, for control.
}

