% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnectance.R
\name{getConnectance}
\alias{getConnectance}
\title{Compute connectance of an interaction matrix.}
\usage{
getConnectance(A)
}
\arguments{
\item{A}{interaction matrix}
}
\value{
the connectance
}
\description{
The connectance is defined as E/N<U+02C6>2-N, where E is the number
of realized arcs (the number of non-zero entries in the interaction matrix) 
and N*(N-1) the number of possible arcs. 
The diagonal (self-arcs) is excluded.
}
\examples{
getConnectance(adjustc(generateA(N=10),c=0.5))
}

