% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPep.R
\name{getPep}
\alias{getPep}
\title{Compute the positive edge percentage of the interaction matrix}
\usage{
getPep(A)
}
\arguments{
\item{the}{interaction matrix (values on the diagonal should be negative)}
}
\value{
the positive edge percentage
}
\description{
Compute the positive edge percentage of the interaction matrix
}

