% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyNoisetypes.R
\name{identifyNoisetypes}
\alias{identifyNoisetypes}
\title{Identify noise types in a count matrix row-wise}
\usage{
identifyNoisetypes(x, pval.threshold = 0.05, abund.threshold = 10)
}
\arguments{
\item{x}{a count matrix}

\item{pval.threshold}{significance threshold for periodogram powerlaw goodness of fit}

\item{abund.threshold}{minimum count sum per row}
}
\value{
S3 noisetypes object
}
\description{
Row count sums need to be above the given abundance threshold and
the spectrum needs to scale significantly with the frequency below p-value 0.05
in log-log scale.
The periodogram is computed with spectrum.
}
\examples{
N=10
ricker.out=ricker(N,generateA(N),K=rep(0.01,N))
identify.noisetypes(ricker.out)
}

