% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{limits}
\alias{limits}
\title{LIMITS implementation
LIMITS is an algorithm developed by Fisher & Mehta to estimate the interaction matrix assuming a Ricker model.}
\usage{
limits(x, bagging.iter = 100, verbose = FALSE)
}
\arguments{
\item{x}{time series with taxa as rows and time points as columns}

\item{bagging.iter}{the number of iterations used for bagging}

\item{verbose}{print which taxon LIMITS is processing}
}
\value{
the estimated interaction matrix
}
\description{
LIMITS implementation
LIMITS is an algorithm developed by Fisher & Mehta to estimate the interaction matrix assuming a Ricker model.
}
\references{
Fisher & Mehta (2014). Identifying Keystone Species in the Human Gut Microbiome from Metagenomic Timeseries using Sparse Linear Regression. PLoS One 9:e102451
}

