% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisetypes.R
\name{noisetypes}
\alias{noisetypes}
\title{Constructor for S3 noisetypes class}
\usage{
noisetypes(white = 0, pink = 0, brown = 0, black = 0)
}
\arguments{
\item{white}{indices of matrix rows with white noise}

\item{pink}{indices of matrix rows with pink noise}

\item{brown}{indices of matrix rows with brown noise}

\item{black}{indices of matrix rows with black noise}
}
\value{
noisetypes class
}
\description{
Construct a noisetypes object.
}
