% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{perturbToBinary}
\alias{perturbToBinary}
\title{Given a perturbation object, extract a binary vector}
\usage{
perturbToBinary(perturb, l = NA, returnCol = FALSE,
  defaultColor = rgb(0, 1, 0, 0.5), perturbColor = rgb(1, 0, 0, 0.5))
}
\arguments{
\item{perturb}{perturbation object}

\item{l}{length of the time series}

\item{returnCol}{return colors instead of FALSE/TRUE values}

\item{defaultColor}{the color used for non-perturbed time points}

\item{perturbColor}{the color used for perturbed time points}
}
\description{
Non-perturbed time points are coded as FALSE, perturbed ones as TRUE.
If returnCol is TRUE, return instead a color vector with the defaultColor for non-perturbed
time points and the perturbColor for perturbed time points.
}
