% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.noisetypes.R
\name{plot.noisetypes}
\alias{plot.noisetypes}
\title{Do a barplot of the noise type counts}
\usage{
\method{plot}{noisetypes}(noisetypes)
}
\arguments{
\item{noisetypes}{the noisetypes object}
}
\description{
Do a barplot of the noise type counts
}
