% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA.R
\name{plotA}
\alias{plotA}
\title{Plot an interaction matrix.}
\usage{
plotA(A, method = "image", header = "", display.weight = FALSE,
  show.grid = FALSE)
}
\arguments{
\item{A}{interaction matrix}

\item{method}{image or corrplot (corrplot requires the corrplot package being loaded first, image is therefore default)}

\item{header}{the title of the plot}

\item{display.weight}{color-code the interaction strength by varying the red/green hues (only for corr.plot)}

\item{show.grid}{display the grid (only for corrplot)}
}
\description{
For plotting, the interaction strengths are scaled to lie between -1 and 1. Negative values
are plotted in red, positive in green.
If corrplot is available, it can be used for plotting.
}
\examples{
plotA(generateA(20,c=0.1))
}

