% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNoisetypes.R
\name{plotNoisetypes}
\alias{plotNoisetypes}
\title{Do a barplot of the noise types}
\usage{
plotNoisetypes(noisetypes)
}
\arguments{
\item{noisetypes}{the noisetypes object}
}
\description{
Plot the count for each noise type in a barplot.
}
