% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerspec.R
\name{powerspec}
\alias{powerspec}
\title{Report slope of periodogram in log-log scale}
\usage{
powerspec(v, plot = FALSE, header = "", col = "blue")
}
\arguments{
\item{v}{time series vector}

\item{plot}{plot the periodogram with the power law in log-scale}

\item{header}{header string}

\item{col}{color used in periodogram if plot is true}
}
\value{
return the slope, p-value, adjusted R2, log frequencies and log spectra
}
\description{
The function uses stats::spectrum to compute the periodogram. It also reports the significance
and goodness of fit of the power law.
}
