% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propertiesA.R
\name{propertiesA}
\alias{propertiesA}
\title{Print properties of the interaction matrix}
\usage{
propertiesA(A, plot.degree = FALSE)
}
\arguments{
\item{A}{interaction matrix}

\item{plot.degree}{plot the degree distribution}
}
\description{
The interaction matrix is converted into a graph and a number
of graph properties (modularity, average clustering coefficient, average path length) are computed using igraph functions.
}

