% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLowAbundance.R
\name{removeLowAbundance}
\alias{removeLowAbundance}
\title{Remove lowest abundance species}
\usage{
removeLowAbundance(dataset, removeN)
}
\arguments{
\item{dataset}{Abundance dataset}

\item{removeN}{Number of species to remove}
}
\value{
Dataset without N lowest abundance species
}
\description{
Removes a number of lowest total abundace species from a dataset generated with generateDataSet.
Note that the function uses the normalized dataset to decide which species to remove, but returns the absolute dataset.
}
