% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\name{ricker}
\alias{ricker}
\title{Generate time series with the Ricker model}
\usage{
ricker(N, A, K = rep(0.1, N), y = runif(N), sigma = 0.05,
  K.trend = NA, tend = 100, tskip = 0, explosion.bound = 10^8,
  perturb = NULL)
}
\arguments{
\item{N}{number of species}

\item{A}{interaction matrix}

\item{K}{a vector of carrying capacities}

\item{y}{initial abundances}

\item{sigma}{noise level, if set to a non-positive value, no noise is added}

\item{K.trend}{a vector of positive or negative carrying capacity percentages specified between 0 and 1}

\item{tend}{number of time points}

\item{tskip}{number of initial time points to skip (to skip the transient)}

\item{explosion.bound}{boundary for explosion}

\item{perturb}{a perturbation object}
}
\value{
a matrix with species abundances as rows and time points as columns
}
\description{
The Ricker model is a discrete version of the generalized Lotka-Volterra model and
is implemented here as proposed by Fisher and Mehta in PLoS ONE 2014.
}
\details{
If an explosion occurs, -1 is returned. An explosion
is reported if any species crosses the explosion bound. To avoid explosions, decrease the number of
interactions in the interaction matrix and/or increase the number of negative interactions in the
interaction matrix. K.trend contains carrying capacity percentages (from the original carrying capacity) that are added
to each species' current carrying capacity in each time step. K.trend allows to simulate an environmental trend.
Negative carrying capacities are set to zero. If both a perturbation and K.trend are specified, the perturbation is carried out first.
}
\examples{
tsplot(ricker(10,generateA(10),K=rep(0.01,10)),header="ricker")
per=perturbation(times=seq(10,100,10),durations=rep(1,10),numberchanges=c(0.1,rep(0,9)))
tsplot(ricker(10,generateA(10),K=rep(0.01,10),perturb=per))
}
\references{
Fisher & Mehta (2014). Identifying Keystone Species in the Human Gut Microbiome from Metagenomic Timeseries using Sparse Linear Regression. PLoS One 9:e102451
}
\seealso{
\code{\link{glv}} for the generalized Lotka Volterra model
}
