% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtime.R
\docType{package}
\name{seqtime}
\alias{seqtime}
\alias{seqtime-package}
\title{Time Series Analysis of Sequencing Data}
\description{
Provides functions to analyse sequencing data time series and to simulate community dynamics.
}

