% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNeg.R
\name{setNeg}
\alias{setNeg}
\title{Set given percentage of interactions to negative values.}
\usage{
setNeg(A, perc = 50, negval = -1, symmetric = TRUE)
}
\arguments{
\item{A}{interaction matrix}

\item{perc}{percentage of interactions to be modified}

\item{negval}{the negative value to be set}

\item{only}{introduce symmetric interactions}
}
\value{
the modified interaction matrix
}
\description{
The number of interactions equals the number of non-zero entries in the interaction matrix.
If symmetric is true, both matrix triangles will be set to a negative value (corresponding
to competition), else only one triangle will be set to a negative value (corresponding
to parasitism, predation or amensalism). If symmetric is true, the number of interactions
is counted only in one triangle of the interaction matrix.
}

