% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHubbell.R
\name{simHubbell}
\alias{simHubbell}
\title{Simulate species abundances with the basic Hubbell model.}
\usage{
simHubbell(N = 50, I = 500, m.vector = runif(N), m = 0.02, d = 10,
  tend = 100)
}
\arguments{
\item{N}{species number}

\item{I}{number of individuals}

\item{m.vector}{species proportions in the metacommunity}

\item{m}{immigration rate (probability that a dead individuum is replaced by an individuum from the metacommunity)}

\item{d}{number of deaths at each time step}

\item{tend}{number of time points (i.e. the number of generations)}
}
\value{
a matrix with species abundances as rows and time points as columns
}
\description{
Simulate species abundances with the basic Hubbell model.
}

