% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuntb.R
\name{simUntb}
\alias{simUntb}
\title{Run the Unified Neutral Theory of Biodiversity (UNTB) model}
\usage{
simUntb(N, y = rep(1, N), m = 0.02, tskip = 0, tend = 5000)
}
\arguments{
\item{N}{number of species}

\item{y}{initial abundances}

\item{m}{migration probability between 0 and 1}

\item{tskip}{number of initial time points to be skipped when returning the result (to avoid the transient)}

\item{tend}{number of time points (i.e. the number of generations)}
}
\value{
a matrix with species abundances as rows and time points as columns
}
\description{
This function just provides a wrapper around the untb function in the untb package.
}
\examples{
N=50
y=generateAbundances(N, mode=5)
tsplot(simUntb(N,y=y,tend=1000))
}
\seealso{
\code{\link{ricker}} for the Ricker model and \code{\link{glv}} for the generalized Lotka Volterra model
}
