% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor.R
\name{taylor}
\alias{taylor}
\title{Plot relationship between row mean and row variance}
\usage{
taylor(x, type = "taylor", pseudo = 0, col = "black", header = "",
  plot = TRUE)
}
\arguments{
\item{x}{a matrix}

\item{type}{the type of plot to do: mean.var (mean vs variance), boxplot (row-wise), taylor (powerlaw fitted to mean vs variance)}

\item{pseudo}{add a pseudo count to deal with zeros in log-log plot (for type=taylor)}

\item{col}{the color of the dots}

\item{header}{header string}

\item{plot}{whether to do the plot}
}
\value{
for type taylor, the slope, p-value and adjusted R2 of the Taylor law are returned (slope, pval, adjR2)
}
\description{
The power law relationship between variance and mean is known as Taylor's law, which is
defined as: \eqn{var(Y) = a*mean(Y)^b}. We can obtain the power as the slope in log-scale: \eqn{log(var(Y)) = log(a)+b*log(mean(Y))}
}
\references{
L.R. Taylor (1961). Aggregation, variance and the mean. Nature 189, 732-735.
}
