% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempDecay.R
\name{tempDecay}
\alias{tempDecay}
\title{Do a time-decay plot of a taxon matrix}
\usage{
tempDecay(x, time = c(1:ncol(x)), dissim = "bray", header = "",
  units = "")
}
\arguments{
\item{x}{a taxon matrix with rows representing taxa and columns samples}

\item{time}{vector with time steps}

\item{dissim}{sample dissimilarity to use, should be supported by vegdist}
}
\description{
Plot the time against the sample distance.
}

