% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStability.R
\name{testStability}
\alias{testStability}
\title{Stability test for interaction matrix}
\usage{
testStability(A, method = "eigen", K = rep(0.1, N), y = runif(N),
  sigma = 0.01, explosion.bound = 10^4, tend = 100)
}
\arguments{
\item{A}{the interaction matrix to test}

\item{method}{the method to test stability (coyte, eigen, ricker)}

\item{K}{carrying capacities for ricker test}

\item{y}{initial abundances for ricker test}

\item{sigma}{noise term for ricker test}

\item{explosion.bound}{explosion boundary for ricker test}

\item{tend}{end of simulation time for ricker test}
}
\value{
boolean false if unstable and true if stable
}
\description{
Test the stability of an interaction matrix.
The first two methods (coyte and eigen) determine whether the community's
steady state is stable in the sense that the community returns to it after
a small perturbation. The last method tests for explosion in simulations
with Ricker.
}
\details{
Coyte's stability criterium is fulfilled if: max(r_e,r_s) - s < 0,
where s is the average of the diagonal values (the intra-species competition),
r_e is the half-horizontal radius of the eigenvalue ellipse of A
(which equals its Jacobian when assuming that species abundances sum to one) and r_s is
the eigenvalue corresponding to the average row sum.
The more negative max(r_e,r_s) - s is, the quicker the community returns
to steady state.
This criterium assumes that realized inter-species interaction strengths are drawn
from a half normal distribution |N(0,sigma2)| (the absolute of the normal
distribution). Only non-zero, non-diagonal interactions are considered to
compute the mean interaction strength.
The eigen method implements the classical stability criterion, which tests whether all real
parts of the eigenvalues of the interaction (=Jacobian) matrix are smaller than zero.
}
\examples{
A=generateA(N=20)
testStability(A)
}
\references{
Coyte et al. (2015). The ecology of the microbiome: Networks, competition, and stability. Science 350:663-666.
}
