% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDecay.R
\name{timeDecay}
\alias{timeDecay}
\title{Do a time-decay plot of a taxon matrix}
\usage{
timeDecay(x, time = c(1:ncol(x)), dissim = "bray", normtaxa = FALSE,
  logdissim = FALSE, logtime = FALSE, header = "", units = "",
  plot = TRUE)
}
\arguments{
\item{x}{a taxon matrix with rows representing taxa and columns samples}

\item{time}{vector with time steps}

\item{dissim}{sample dissimilarity to use, should be supported by vegdist}

\item{normtaxa}{divide each taxon vector in x by its sum}

\item{logdissim}{take the logarithm of the dissimilarity before fitting a line}

\item{logtime}{take the logarithm of the time interval before fitting a line}

\item{header}{a string to be appended to the plot title (Time decay)}

\item{units}{a string to describe the units of the time points (days, weeks etc)}

\item{plot}{whether to do the plot}
}
\value{
a list with the time intervals, dissimilarity values, intersection, slope, p-value, adjusted R2, dissimilarity measure used and log status of dissimilarity measure
}
\description{
Plot the time against the sample distance.
}
\examples{
data("david_stoolA_otus")
data=rarefyFilter(david_stoolA_otus,min=10000)[[1]]
out.decay=timeDecay(data[,1:50], header="Stool subject A")
}

