% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Plot the time series row-wise}
\usage{
tsplot(x, time.given = FALSE, num = nrow(x), sample.points = c(),
  legend = FALSE, pch = "+", lty = 1, type = "l", header = "")
}
\arguments{
\item{x}{the matrix of time series}

\item{time.given}{if true, then the column names are supposed to hold the time units}

\item{num}{the number of rows to plot (starting from the first row)}

\item{legend}{add a legend}

\item{pch}{plot parameter}

\item{type}{pot parameter}

\item{header}{header string}
}
\description{
Plot the time series row-wise
}
