% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignTaxonLevelsToA.R
\name{assignTaxonLevelsToA}
\alias{assignTaxonLevelsToA}
\title{Assign Taxon Levels}
\usage{
assignTaxonLevelsToA(A, data = NULL, metadata = NULL, lineages, type = "",
  taxon.level = "genus", uniqueNames = FALSE, higherLevelNames = TRUE)
}
\arguments{
\item{A}{an interaction matrix}

\item{data}{(optional) a matrix with taxon abundances, rows are taxa and columns are samples (only needed if type is non-empty)}

\item{metadata}{(optional) a matrix with metadata, samples match samples in OTU table (only needed if type is non-empty)}

\item{lineages}{a matrix with lineages, for the format please see data david_stool_lineages}

\item{type}{if non-empty, process data exactly as in method generateTS, supported: stoola and stoolb}

\item{taxon.level}{the taxon level to be assigned as row and column names, supported: otu, species, genus, family, order, class, phylum}

\item{uniqueNames}{make names unique by appending a counter if needed}

\item{higherLevelNames}{if given level is not known, assign the highest level that is known}
}
\description{
Assign OTU identifiers or higher-level taxon names to an interaction matrix
}
\examples{
data("david_stool_lineages")
# make a random OTU interaction matrix as an example
N=30
A=generateA(N=N,c=0.1)
randomOTUIndices=sample(1:nrow(david_stool_lineages))[1:N]
rownames(A)=david_stool_lineages[randomOTUIndices,1]
colnames(A)=rownames(A)
Aclass=assignTaxonLevelsToA(A,lineages=david_stool_lineages, taxon.level="class")
classnetwork=plotA(Aclass,method="network")
}
