% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTS.R
\name{compareTS}
\alias{compareTS}
\title{Community Time Series Comparison with generateTS}
\usage{
compareTS(input.folder = "", expIds = c(), modif.folder = "",
  modif = "", sliceDef = c(1, NA), epsilon = 0.2, norm = FALSE,
  hurstBins = c(0.5, 0.7, 0.9), maxautocorBins = c(0.3, 0.5, 0.8),
  timeDecaySliceDef = c(1, 50), varEvolSliceDef = c(),
  returnDistribs = FALSE, returnTS = FALSE)
}
\arguments{
\item{input.folder}{the folder where results (settings and time series) of function generateTS are stored.}

\item{expIds}{the experiment identifiers of time series to be considered}

\item{modif.folder}{the folder with time series sub-sets as generated by sliceTS.R or with noisy time series as generated by addNoise.R (settings are read in from input.folder)}

\item{modif}{modification (shared name of time series in modif.folder, sliced for time series sub-sets and pois for time series with Poisson noise added)}

\item{sliceDef}{consider the time series sub-set defined by the given start and end time point for property computation, an end time point of NA means that the entire time series is considered; timeDecaySliceDef refers to the outcome of sliceDef}

\item{epsilon}{allowed deviation from the expected slope for noise type identification}

\item{norm}{normalize time series by dividing each entry in a sample by the sum of its sample; if norm is true, entropy computation is omitted}

\item{hurstBins}{binning thresholds for Hurst exponent (three thresholds required for four bins)}

\item{maxautocorBins}{binning thresholds for maximal autocorrelation (three thresholds required for four bins)}

\item{timeDecaySliceDef}{the time series subset considered to compute the time decay, a vector with the start and the end time point (if the end time point is NA, the entire time series is used)}

\item{varEvolSliceDef}{the time series subset considered to compute the evolution of variance, a vector with the start and the end time point (if the end time point is NA, the entire time series is used)}

\item{returnDistribs}{return distributions at the final time point (no time series properties are computed, cannot be used together with returnTS)}

\item{returnTS}{return the time series (no time series properties are computed, cannot be used together with returnDistribs)}
}
\value{
a table with experiment parameters (algorithm, connectance, sigma, theta and so on) and time series properties (noise types, slope of Taylor's law etc.); if returnDistrib is true, a list with the abundances at the last time point, if returnTS is true, a list with the time series
}
\description{
Compare community time series generated with generateTS.
}
\details{
If R package infotheo is installed, the entropy of time series
is computed.
}
