% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTS.R
\name{compareTS}
\alias{compareTS}
\title{Compare community time series generated with generateTS}
\usage{
compareTS(input.folder = "", expIds = c(), timeDecayInterval = 20)
}
\arguments{
\item{input.folder}{the folder where results of function generateTS are stored.}

\item{expIds}{the experiment identifiers of time series to be considered}

\item{timeDecayInterval}{the interval considered to compute the time decay}
}
\value{
a table with experiment parameters (algorithm, connectance, sigma, theta and so on) and time series properties (noise types, slope of Taylor's law etc.)
}
\description{
Compare community time series generated with generateTS
}

