% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david_stoolA_otus.R
\docType{data}
\name{david_stoolA_otus}
\alias{david_stoolA_otus}
\title{David Stool Sequencing Data for Subject A}
\format{An object of class \code{data.frame} with 5431 rows and 329 columns.}
\usage{
david_stoolA_otus
}
\value{
OTUxSample matrix of size 5431 x 329.
}
\description{
The OTUs sequenced from stool samples of subject A over time.
}
\details{
Subject A traveled from day 71 to day 122 and contracted diarrhea on
days 80 to 85 and 104 to 113.
}
\examples{
data(david_stoolA_otus)
  # sort OTUs by abundance
  sorted <- sort(apply(david_stoolA_otus,1,sum),
                 decreasing = TRUE, index.return = TRUE) 
# display the top 10 most abundant OTUs
tsplot(david_stoolA_otus[sorted$ix[1:10],], type = "l", header = "Stool A") 
}
\references{
David et al. (2014) Host lifestyle affects human microbiota on daily timescales Genome Biology vol. 15 (7):R89
\href{http://www.genomebiology.com/2014/15/7/R89}{Genome Biology}
}
\keyword{data,}
\keyword{datasets}
