% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc}
\alias{doc}
\title{The DOC method by Bashan et al.}
\usage{
doc(x, B = 100, polygons = FALSE, rand = FALSE, lower.conf = 0.03,
  upper.conf = 0.97, null.model = "assembly")
}
\arguments{
\item{x}{rows are taxa, columns are samples}

\item{B}{bootstrap iterations, if set to 0 or below, no bootstraps are carried out}

\item{polygons}{draw a polygon instead of the lower and upper confidence line}

\item{rand}{randomize the data sample-wise before starting the DOC analysis}

\item{lower.conf}{lower limit of the confidence interval}

\item{upper.conf}{upper limit of the confidence interval}

\item{null.model}{the null model to use, permut shuffles x sample-wise, assembly selects for each non-zero taxon one of the values taken across the samples at random}
}
\value{
a list with the overlaps, dissimilarities, lowess smoothed overlaps and dissimilarites and lower and upper confidence intervals
}
\description{
Given  matrix of relative taxon abundances, plot the dissimilarity-overlap curve (DOC) sample-wise.
The DOC method was developed by Bashan and colleagues.
}
\examples{
data("david_stoolA_otus")
data=rarefyFilter(david_stoolA_otus,min=10000)[[1]]
out=doc(data[,1:100],B=10) # apply the DOC method on the first 100 time points
N=50
M=500
metapop=generateAbundances(N=M, mode=5, probabs=TRUE)
ts=simHubbell(N=N, M=M,I=1500,d=N, m.vector=metapop, tskip=500, tend=600)
out2=doc(ts,B=10) # apply the DOC method to the Hubbell time series
}
\references{
A. Bashan et al. (2016). Universality of human microbial dynamics, Nature 534, 259-262.
}

