% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc}
\alias{doc}
\title{Given  matrix of relative taxon abundances, plot the dissimilarity-overlap curve (DOC) sample-wise.}
\usage{
doc(x, B = 100, polygons = FALSE, rand = FALSE, lower.conf = 0.03,
  upper.conf = 0.97, null.model = "assembly")
}
\arguments{
\item{x}{rows are taxa, columns are samples}

\item{B}{bootstrap iterations, if set to 0 or below, no bootstraps are carried out}

\item{polygons}{draw a polygon instead of the lower and upper confidence line}

\item{rand}{randomize the data sample-wise before starting the DOC analysis}

\item{lower.conf}{lower limit of the confidence interval}

\item{upper.conf}{upper limit of the confidence interval}

\item{null.model}{the null model to use, permut shuffles x sample-wise, assembly selects for each non-zero taxon one of the values taken across the samples at random}
}
\value{
a list with the overlaps, dissimilarities, lowess smoothed overlaps and dissimilarites and lower and upper confidence intervals
}
\description{
Given  matrix of relative taxon abundances, plot the dissimilarity-overlap curve (DOC) sample-wise.
}

