% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAbundances.R
\name{generateAbundances}
\alias{generateAbundances}
\title{Generate abundance vector}
\usage{
generateAbundances(N, count = 1000, mode = 1, k = 0.5, probabs = FALSE)
}
\arguments{
\item{N}{number of species}

\item{count}{total number of individuals (ignored in mode 2, 4 and 7)}

\item{mode}{1=each species receives round(count/N) counts, 2=sampled from uniform distribution with 0 as lower and count as upper bound, 3=dominant species takes 95 percent of the counts and all others split the remaining counts equally, 4=counts are sampled from a Poisson distribution with lambda set to count/N, 5=using bstick function from vegan, 6=using geometric series with parameter k, 7=sample from the exponential distribution and scale with count/N}

\item{k}{evenness parameter of geometric series}

\item{probabs}{if true, return probabilities instead of abundances}
}
\value{
species abundances
}
\description{
Generate a vector of species abundances that can be used
as initial species abundance vector in models.
}
\details{
Note: The geometric series (a niche apportionment model) is for instance described in
"Comments about some species abundance patterns: classic, neutral, and niche partitioning models"
by Ferreira and Petrere-Jr., Braz. J. Biol. 2008
}
\examples{
round(generateAbundances(10,mode=6))
}
\seealso{
\code{\link{simCountMat}}
}
