% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataSet.R
\name{generateDataSet}
\alias{generateDataSet}
\title{Generate a dataset}
\usage{
generateDataSet(samples, matrix, env.matrix = NULL, perturb.count = NULL,
  count = 1000, mode = 4)
}
\arguments{
\item{samples}{Number of samples}

\item{matrix}{Interaction matrix (generated from generateA.R)}

\item{env.matrix}{Growth rate changes induced by environment; 1 column per environmental condition. Generate this matrix with envGrowthChanges.R}

\item{perturb.count}{Number of samples per environmental condition. Sum should be equal to total number of samples.}

\item{count}{Total number of individuals in dataset}

\item{mode}{Mode for generateAbundances; default value samples counts from Poisson distribution with lambda count/N}
}
\value{
The abundance dataset
}
\description{
Generate an abundance dataset,
with or without environmental perturbance, by using generalized Lotka-Volterra
}
\examples{
klemm = generateA(N=10, type="klemm", c=0.5)
env = envGrowthChanges(species = 10, env.factors=2, conditions=2, strength=0.5)
dataset = generateDataSet(100, klemm, env.matrix = env, perturb.count = c(50, 50))
}
