% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnectance.R
\name{getConnectance}
\alias{getConnectance}
\title{Connectance}
\usage{
getConnectance(A)
}
\arguments{
\item{A}{an interaction matrix}
}
\value{
the connectance
}
\description{
Compute connectance of an interaction matrix.
}
\details{
The connectance is defined as \eqn{c=E/(N*(N-1))}, where \eqn{E} is the number
of realized arcs (the number of non-zero entries in the interaction matrix)
and \eqn{N*(N-1)} the number of possible arcs.
The diagonal (self-arcs) is excluded.
}
\examples{
  A <- cbind(c(-1,0,1),c(0,-1,0),c(-1,0,-1))
  x <- getConnectance(A)
}
