% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPep.R
\name{getPep}
\alias{getPep}
\title{Compute the positive edge percentage of the interaction matrix}
\usage{
getPep(A)
}
\arguments{
\item{the}{interaction matrix}
}
\value{
the positive edge percentage
}
\description{
Values on the diagonal are ignored.
}
\examples{
A=cbind(c(-1,0,1),c(0,-1,0),c(-1,0,-1))
getPep(A)
}
