% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPep.R
\name{getPep}
\alias{getPep}
\title{Positive Edge Percentage}
\usage{
getPep(A)
}
\arguments{
\item{A}{the interaction matrix}
}
\value{
the positive edge percentage
}
\description{
Compute the positive edge percentage of an interaction matrix.
}
\details{
Values on the diagonal are ignored. More precisely, the positive arc percentage is computed, e.g.
the percentage of positive entries in the interaction matrix.
}
\examples{
  A <- cbind(c(-1,0,1),c(0,-1,0),c(-1,0,-1))
  x <- getPep(A)
}
