% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonomy.R
\name{getTaxonomy}
\alias{getTaxonomy}
\title{Get the taxonomy given OTU names and lineage information}
\usage{
getTaxonomy(selected = c(), lineages, level = "class")
}
\arguments{
\item{selected}{a character vector of selected OTU identifiers}

\item{lineages}{a lineage table}

\item{level}{the taxonomic level (domain, phylum, class, order, family, genus or species)}
}
\value{
the taxonomy of the OTUs
}
\examples{
data(david_stoolA_otus)
data(david_stool_lineages)
sorted=sort(apply(david_stoolA_otus,1,sum),decreasing=TRUE)[1:10]
getTaxonomy(names(sorted),david_stool_lineages,level="family")
}
