% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glv.R
\name{glv}
\alias{glv}
\title{Generalized Lotka-Volterra model}
\usage{
glv(N = 4, A, b = runif(N), y = runif(N), tstart = 0, tend = 100,
  tstep = 0.1)
}
\arguments{
\item{N}{species number}

\item{A}{interaction matrix}

\item{b}{growth rates}

\item{y}{initial abundances}

\item{tstart}{initial time point}

\item{tend}{final time point}

\item{tstep}{time step}
}
\value{
a matrix with species abundances as rows and time points as columns, column names give time points
}
\description{
Generalized Lotka-Volterra model
}
\examples{
tsplot(glv(N=4,generateA(4)),header="gLV",type="l")
}
\seealso{
\code{\link{ricker}} for the Ricker model
}

