% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyNoisetypes.R
\name{identifyNoisetypes}
\alias{identifyNoisetypes}
\title{Identify Noise Types}
\usage{
identifyNoisetypes(x, epsilon = 0.2, pval.threshold = 0.05,
  permut = FALSE, abund.threshold = 0)
}
\arguments{
\item{x}{a matrix with objects as rows and time points as columns}

\item{epsilon}{allowed deviation from the expected slope of 0 for white noise, -1 for pink noise and -2 for brown noise (all rows with a slope below -3 are classified as having black noise)}

\item{pval.threshold}{significance threshold for periodogram powerlaw goodness of fit}

\item{permut}{permute time points before computing noise types}

\item{abund.threshold}{minimum sum per row}
}
\value{
S3 noisetypes object
}
\description{
Identify noise types in a matrix row-wise.
}
\details{
Row sums need to be above the given abundance threshold and the spectral density needs to scale significantly with the frequency (p-value below 0.05) in log-log scale. The periodogram is computed with \code{\link{spectrum}} from the \pkg{stats} package. The function returns a noisetypes object, which groups matrix row indices by noise type.
}
\examples{
N <- 10
ricker.out <- ricker(N,generateA(N),K=rep(0.01,N))
noise <- identifyNoisetypes(ricker.out, abund.threshold=0)
plot(noisetypes)

}
