% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyNoisetypes.R
\name{identifyNoisetypes}
\alias{identifyNoisetypes}
\title{Identify noise types in a matrix row-wise}
\usage{
identifyNoisetypes(x, pval.threshold = 0.05, abund.threshold = 10)
}
\arguments{
\item{x}{a matrix with objects as rows and time points as columns}

\item{pval.threshold}{significance threshold for periodogram powerlaw goodness of fit}

\item{abund.threshold}{minimum count sum per row}
}
\value{
S3 noisetypes object
}
\description{
Row sums need to be above the given abundance threshold and the spectral density
needs to scale significantly with the frequency (p-value below 0.05) in log-log scale.
The periodogram is computed with stats::spectrum.
The function returns a noisetypes object, which groups matrix row indices
by noise type.
}
\examples{
N=10
ricker.out=ricker(N,generateA(N),K=rep(0.01,N))
noisetypes=identifyNoisetypes(ricker.out)
plot(ricker.out[noisetypes$brown[1],], main=paste("Simulated OTU",noisetypes$brown[1]),ylab="Abundance")
}

