% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyA.R
\name{modifyA}
\alias{modifyA}
\title{Modify the interaction matrix}
\usage{
modifyA(A, mode = "adjustc", strength = "binary", minstrength = 0.1,
  c = 0.2, perc = 50, symmetric = FALSE)
}
\arguments{
\item{A}{the interaction matrix}

\item{mode}{modification mode, adjustc (adjust connectance to reach target connectance), negpercent (set the specified percentage of negative edges), tweakstable (randomly add negative edges until matrix is stable or fully connected)}

\item{strength}{interaction strength, binary (0/1) or uniform (sampled from uniform distribution from minstrength to 1)}

\item{minstrength}{minimum interaction strength for uniform mode (maximum is 1)}

\item{c}{the target connectance (only for mode adjustc)}

\item{perc}{negative edge percentage (only for mode negpercent)}

\item{symmetric}{only introduce symmetric negative interactions (only for mode negpercent)}
}
\value{
the modified interaction matrix
}
\description{
Mode negpercent: The number of interactions equals the number of non-zero entries in the interaction matrix.
If symmetric is true, both matrix triangles will be set to a negative value (corresponding
to competition), else only one triangle will be set to a negative value (corresponding
to parasitism, predation or amensalism). If symmetric is true, the number of interactions
is counted only in one triangle of the interaction matrix.
}

