% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisetypes.R
\name{noisetypes}
\alias{noisetypes}
\title{Constructor for S3 noisetypes class}
\usage{
noisetypes(white = 0, pink = 0, brown = 0, black = 0)
}
\arguments{
\item{white}{percentage of white noise rows in matrix}

\item{pink}{percentage of pink noise rows in matrix}

\item{brown}{percentage of brown noise rows in matrix}

\item{black}{percentage of black noise rows in matrix}
}
\value{
noisetypes class
}
\description{
Construct a noisetypes object.
}
