% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA.R
\name{plotA}
\alias{plotA}
\title{Plot an interaction matrix.}
\usage{
plotA(A, method = "image", header = "", scale.weight = FALSE,
  original = FALSE)
}
\arguments{
\item{A}{interaction matrix}

\item{method}{image or ggplot (ggplot requires ggplot2 and reshape2, image is therefore default)}

\item{header}{the title of the plot}

\item{scale.weight}{scale interaction strengths between -1 and 1}

\item{original}{plot original values}
}
\description{
By default, the interaction strengths are set to -1 or 1. Negative values
are plotted in red, positive in green. If scale.weight is true, the interaction
strengths are scaled to lie in the range of [-1,1]. The option original suppresses
any modification of the interaction strengths. If interaction strengths are scaled or the original
ones are used, the method ggplot is recommended, since it adds a color legend.
Method ggplot requires ggplot2 and reshape2.
}
\examples{
plotA(generateA(20,c=0.1))
}

