% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLimitsQuality.R
\name{plotLimitsQuality}
\alias{plotLimitsQuality}
\title{Quality plot for interaction matrices estimated with LIMITS}
\usage{
plotLimitsQuality(oriTS, A, A.ori = matrix())
}
\arguments{
\item{oriTS}{the original time series, with taxa as rows and time points as columns}

\item{A}{the estimated interaction matrix}

\item{A.ori}{the original interaction matrix (optional)}

\item{sigma}{noise term for time series generation with ricker}

\item{explosion.bound}{explosion boundary for time series generation with ricker}
}
\value{
a list with the taxon numbers considered (taxonnum), the mean correlation of observed and predicted time series from lag 0 to lag 4 (meancrosscor and meancrosscor1 to meancrosscor4) and the mean autocorrelations up to lag 4 (meanautocor1 to meanautocor4)
}
\description{
To avoid explosions, the estimated interaction matrix is modified such
that all its eigen values are negative. The predicted time series is
obtained step-wise, by computing the current abundances from the
original abundances of the preceding time point and the given
estimated interaction matrix with (noise-free) Ricker. Carrying capacities are
estimated as mean abundances. If the original interaction matrix is provided,
the mean correlation between estimated and original interaction matrix is computed.
}

