% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\name{ricker}
\alias{ricker}
\title{Generate time series with the Ricker model}
\usage{
ricker(N, A, K = rep(0.1, N), y = runif(N), sigma = 0.05, tend = 100,
  tskip = 0, explosion.bound = 10^8)
}
\arguments{
\item{N}{number of species}

\item{A}{interaction matrix}

\item{K}{carrying capacity}

\item{y}{initial abundances}

\item{sigma}{noise level, if set to a non-positive value, no noise is added}

\item{tend}{number of time points}

\item{tskip}{number of initial time points to skip (to skip the transient)}

\item{explosion.bound}{boundary for explosion}
}
\value{
a matrix with species abundances as rows and time points as columns
}
\description{
The Ricker model is a discrete version of the generalized Lotka-Volterra model and
is implemented here as proposed by Fisher and Mehta in PLoS ONE 2014.
}
\details{
If an explosion occurs, -1 is returned. An explosion
is reported if any species crosses the explosion bound.
}
\examples{
tsplot(ricker(10,generateA(10),K=rep(0.01,10)),type="l", header="ricker")
}
\references{
Fisher & Mehta (2014). Identifying Keystone Species in the Human Gut Microbiome from Metagenomic Timeseries using Sparse Linear Regression. PLoS One 9:e102451
}
\seealso{
\code{\link{glv}} for the generalized Lotka Volterra model
}
