% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDecay.R
\name{simDecay}
\alias{simDecay}
\title{Plot sample-wise taxon dissimilarities against the abundance difference of selected taxa}
\usage{
simDecay(x, taxa = c(1), dissim = "bray", logdissim = FALSE)
}
\arguments{
\item{x}{taxon-sample matrix}

\item{taxa}{vector of taxon row indices or row names present in x}

\item{logdissim}{take the logarithm of the dissimilarity before fitting a line}
}
\description{
Plot sample-wise taxon dissimilarities against the abundance difference of selected taxa
}
\examples{
data=rarefyFilter(david_stoolA_otus,min=10000)
# Faecalibacterium OTUs
taxa=c("OTU_72853","OTU_119271")
out=simDecay(data,taxa)
}

