% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDecay.R
\name{simDecay}
\alias{simDecay}
\title{Plot community similarity decay against selected taxa or metadata}
\usage{
simDecay(x, taxa = c(), metadata = c(), metadataName = "",
  samplegroups = rep(1, ncol(x)), dissim = "bray", normtaxa = FALSE,
  logdissim = FALSE, intragroupsOnly = FALSE, header = "",
  groupName = "", inversePlot = FALSE)
}
\arguments{
\item{x}{taxon-sample matrix with taxa as rows and samples as columns}

\item{taxa}{vector of taxon row indices or row names present in x}

\item{metadata}{vector with values in the same order as samples in the taxon matrix}

\item{metadataName}{name of the metadata vector (e.g. age)}

\item{samplegroups}{supposed to assign an integer to each sample, with integers ranging from 1 to group number. If provided, intra-group dissimilarity dots are colored by sample group}

\item{dissim}{dissimilarity measure to use, a measure supported by vegan's vegdist function}

\item{normtaxa}{divide each taxon vector in x by its sum}

\item{logdissim}{take the logarithm of the dissimilarity before fitting a line}

\item{intragroupsOnly}{only take sample pairs within the same group into account for dissimilarity computation (groups are labeled in the plot by their identifier)}

\item{header}{a string to be appended to the plot title (Dissimilarity change)}

\item{groupName}{the name for the groups displayed in the legend (defaults to group)}

\item{inversePlot}{plot community dissimilarity on the x-axis instead of on the y-axis}
}
\description{
All pair-wise community dissimilarities are plotted
against the corresponding abundance difference of selected taxa or metadata.
}
\examples{
N=50
data("david_stoolA_otus")
rar=rarefyFilter(david_stoolA_otus,min=10000)
data=rar[[1]]
# Faecalibacterium OTUs
taxa=c("OTU_72853","OTU_119271")
out1=simDecay(data[,1:N],taxa)
data("david_stoolA_metadata")
days=david_stoolA_metadata[1,rar[[2]]] # only keep samples that made it through rarefaction
out2=simDecay(data[,1:N],metadata=days[1:N],metadataName=rownames(david_stoolA_metadata)[1])

}
