% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNoiseMat.R
\name{simNoiseMat}
\alias{simNoiseMat}
\title{Simulate a matrix composed of rows with different noise types.}
\usage{
simNoiseMat(samples = 100, noisetypes = list(white = 2, brown = 3, pink =
  5), brown.type = "bm", mean = 5, sd = 1, c = 1, r = 2,
  stayDead = FALSE)
}
\arguments{
\item{samples}{the number of samples to be drawn}

\item{noisetypes}{a list specifying the number of rows to be generated for each noise type}

\item{brown.type}{the algorithm to generate brown noise, options: bm (Brown motion) and gbm (geometrical Brown motion; requires sde package)}

\item{mean}{parameter for white and brown noise (type bm)}

\item{sd}{parameter for white and brown noise (type bm)}

\item{c}{constant added to pink noise}

\item{r}{parameter for brown noise (type gbm)}

\item{stayDead}{once a time series hits zero or below, it stays at zero}
}
\value{
a matrix where each row represents a distribution following a specified noise type
}
\description{
White noise is generated with rnorm and brown noise with cumsum(rnorm).
Pink noise is generated by default with the voss1d function from the Voss package.
Brown noise can optionally be generated with geometric Brown motion via the 
GBM function in the sde package. Option stayDead allows simulating extinction 
events; i.e. once a value of zero or below has been hit, the row only takes zero values.
}
\examples{
# plot the power spectrum of pink noise
powerspec(simNoiseMat(samples=100,noisetypes=list(white=0,brown=0,pink=1))[1,],plot.log=TRUE)
# plot a matrix with mixed noise types
mat=simNoiseMat(samples=300,noisetypes=list(white=0,brown=20,pink=30))
tsplot(mat,type="l", header="simulated noise")
# Taylor law of brown noise
taylor(simNoiseMat(samples=300,noisetypes=list(white=0,brown=50,pink=0)),type="taylor")
# Check generation of pink noise
identifyNoisetypes(simNoiseMat(samples=500,noisetypes=list(white=0,brown=0,pink=10)))
}

