% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliceTS.R
\name{sliceTS}
\alias{sliceTS}
\title{Slice time series}
\usage{
sliceTS(slices = NULL, slice.def.path = "", input.folder = "",
  output.folder = "", expIds = c())
}
\arguments{
\item{slices}{a matrix with rows and 2 columns, where each row codes the time series identifier, the first column the start and the second the end time point; NA means until end of the time series}

\item{slice.def.path}{path to a slice definition file, which has two tab-delimited columns for the start and end time points, respectively; NA means until end of the time series}

\item{input.folder}{location of time series and settings sub folders}

\item{output.folder}{folder in which all slices go (no sub-folders)}

\item{expIds}{set of experiment identifiers to process}
}
\description{
Store slices of previously generated time series in the input folder
and store them in the output folder.
}
\details{
The input folder is expected to contain the folder structure as generated by
generateTS.R (timeseries subfolder containing experiment subfolders).
Sliced time series are stored in the output folder without subfolders.
}
