% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor.R
\name{taylor}
\alias{taylor}
\title{Plot relationship between row mean and row variance}
\usage{
taylor(x, type = "boxplot", pseudo = 0, col = "black", header = "")
}
\arguments{
\item{x}{a matrix}

\item{type}{the type of plot to do: mean.var (mean vs variance), boxplot (row-wise), taylor (powerlaw fitted to mean vs variance)}

\item{pseudo}{add a pseudo count to deal with zeros in log-log plot (for type=taylor)}

\item{col}{the color of the dots}

\item{header}{header string}
}
\description{
The power law of mean and variance is also known as Taylor's law.
Taylor's law: var(Y) = a*mean(Y)^b, in log-scale: log(var(Y)) = log(a)+b*log(mean(Y))
The slope b of Taylor's law is related to the Hurst exponent by: H=b/2 (Kendal 2013).
}
\references{
Kendal, Journal of Basic and Applied Physics 2013, vol 2, iss 2 pp. 40-49 (eq 32)
}

