% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDiagnostic.R
\name{tsDiagnostic}
\alias{tsDiagnostic}
\title{Diagnostic Plot for Community Simulation}
\usage{
tsDiagnostic(x, type = "ind", spec.index = 1, spec.meta = NA)
}
\arguments{
\item{x}{community time series, rows are species and columns are time points}

\item{type}{plot time vs number of species (spec), number of individuals (ind) or the distribution of the species with given index together with its proportion in the metacommunity, if provided}

\item{spec.index}{row index of selected species (required for type distrib)}

\item{spec.meta}{proportion of selected species in the meta-community (optional)}
}
\description{
Diagnostic plots for the outcome of a community simulation.
}
\examples{
N=50
m.vector=generateAbundances(N,mode=5,probabs=TRUE)
outH=simHubbell(M=N,N=N,I=1000,m.vector=m.vector,tend=1000)
tsDiagnostic(outH,type="distrib",spec.index = 4,spec.meta = m.vector[4])
tsDiagnostic(outH,type="spec")
tsDiagnostic(outH,type="ind")
}
