% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDiagnostic.R
\name{tsDiagnostic}
\alias{tsDiagnostic}
\title{Diagnostics for Community Simulation}
\usage{
tsDiagnostic(x, mode = "ind", plot = TRUE, spec.index = 1,
  spec.meta = NA, threshold = 0.01, K = NA, ...)
}
\arguments{
\item{x}{community time series, rows are species and columns are time points}

\item{mode}{plot time vs number of species (spec), number of individuals/sum of abundance (ind) or the distribution of the species with given index (distrib)}

\item{plot}{plot the diagnostic plot of selected mode}

\item{spec.index}{row index of selected species (optional, required for mode distrib)}

\item{spec.meta}{proportion of selected species in the meta-community (optional, if provided it is added in plot mode distrib)}

\item{threshold}{threshold below which species are considered dead}

\item{K}{carrying capacity, in case of Hubbell and SOI, the number of individuals (optional, if provided it is shown in plot mode ind)}

\item{\\dots}{Additional arguments passed to plot()}
}
\value{
depending on the mode, the function returns the number of surviving species over time, the number of individuals over time or the species distribution at the given time point
}
\description{
Diagnostics for the outcome of a community simulation.
Reports the number of species that went extinct at the end of the simulation
and the number and proportion of zeros in the time series. In addition,
it offers diagnostic plots.
}
\examples{
\dontrun{
N=50
m.vector=generateAbundances(N,mode=5,probabs=TRUE)
outH=simHubbell(M=N,N=N,I=1000,m.vector=m.vector,tend=1000)
diagnostics=tsDiagnostic(outH,mode="distrib",spec.index = 4,spec.meta = m.vector[4])
diagnostics=tsDiagnostic(outH,mode="spec")
diagnostics=tsDiagnostic(outH,mode="ind")
}
}
