% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDiagnostic.R
\name{tsDiagnostic}
\alias{tsDiagnostic}
\title{Diagnostics for Community Simulation}
\usage{
tsDiagnostic(x, type = "ind", plot = TRUE, spec.index = 1,
  spec.meta = NA)
}
\arguments{
\item{x}{community time series, rows are species and columns are time points}

\item{type}{plot time vs number of species (spec), number of individuals/sum of abundance (ind) or the distribution of the species with given index (distrib)}

\item{plot}{plot the diagnostic plot of selected type}

\item{spec.index}{row index of selected species (optional, required for type distrib)}

\item{spec.meta}{proportion of selected species in the meta-community (optional, if provided it is added in plot type distrib)}
}
\description{
Diagnostics for the outcome of a community simulation.
Reports the number of species that went extinct at the end of the simulation
and the number and proportion of zeros in the time series. In addition,
it offers diagnostic plots.
}
\examples{
\dontrun{
N=50
m.vector=generateAbundances(N,mode=5,probabs=TRUE)
outH=simHubbell(M=N,N=N,I=1000,m.vector=m.vector,tend=1000)
tsDiagnostic(outH,type="distrib",spec.index = 4,spec.meta = m.vector[4])
tsDiagnostic(outH,type="spec")
tsDiagnostic(outH,type="ind")
}
}
