% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsJumpStats.R
\name{tsJumpStats}
\alias{tsJumpStats}
\title{Compute statistics on jumps through community space}
\usage{
tsJumpStats(x, ordinate = TRUE, distance = "bray",
  ordinate.distance = "bray", dimensions = c(1, 2), plot = FALSE,
  header = "", perturb = NULL)
}
\arguments{
\item{x}{a community time series matrix, with rows as taxa and columns as time points}

\item{ordinate}{if TRUE, compute jumps in ordination plot}

\item{distance}{if ordinate is FALSE, the distance used to compute jumps}

\item{ordinate.distance}{the distance used for ordination}

\item{dimensions}{if ordinate is TRUE, the principal components considered}

\item{plot}{if TRUE, a plot of the jumps is made}

\item{header}{text attached to plot title}

\item{perturb}{a perturbation object, if provided, the plot is colored accordingly}
}
\description{
Consecutive samples representing time points in an ordination plot can be interpreted
as vectors, which have a length and an angle. If ordinate is true, these lengths and angles
are computed. If ordinate is false, dissimilarities between consecutive time points are computed.
}
\examples{
\dontrun{
data(david_stoolA_otus)
data(david_stoolA_metadata)
rarefaction=rarefyFilter(david_stoolA_otus,min = 10000)
rarefiedA=rarefaction$rar
days=david_stoolA_metadata[1,rarefaction$colindices]
interpA=interpolate(rarefiedA,time.vector=days,interval=1,method="stineman")
interpA[interpA<0]=0
perturbA=perturbation(times=c(80,105), durations=c(5,9))
res=tsJumpStats(interpA, plot=TRUE, perturb=perturbA, header="in stool A")
out=powerspec(res$lengths,plot=TRUE)
}
}
