% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsJumpStats.R
\name{tsJumpStats}
\alias{tsJumpStats}
\title{Compute statistics on jumps through community space}
\usage{
tsJumpStats(x, time.given = FALSE, time.unit = "days", ordinate = TRUE,
  distance = "bray", dimensions = c(1, 2), plot = FALSE,
  plot.type = "jumps", header = "", subsample = FALSE, perturb = NULL)
}
\arguments{
\item{x}{a community time series matrix, with rows as taxa and columns as time points}

\item{time.given}{sample names provide time steps, only needed for plotting and plot.type jumps}

\item{time.unit}{unit of time, only needed for plotting and plot.type jumps}

\item{ordinate}{if TRUE, compute jumps in ordination plot}

\item{distance}{the distance (or dissimilarity) used to compute jumps directly or to compute the ordination}

\item{dimensions}{if ordinate is TRUE, the principal components considered}

\item{plot}{if TRUE, a plot of the jumps is made}

\item{plot.type}{jumps: plot jumps as bars in the order of occurrence; box: a box plot of jump lengths in non-perturbed vs perturbed communities. This is effectively a visualization of beta-diversity, which is however not computed between all pair-wise but only between consecutive samples; hist: a histogram of jump lengths for perturbed and non-perturbed sample pairs (box and hist only possible if perturbation object is provided)}

\item{header}{text to be used as plot title instead of default text}

\item{subsample}{for plot.type hist or box: subsample larger jump vector randomly down to smaller one, so that there are as many perturbed as non-perturbed jumps}

\item{perturb}{a perturbation object, if provided, the plot is colored accordingly}
}
\value{
jump lengths (i.e. dissimilarities of community composition at consecutive time points) and, in case ordinate is true, angles are returned
}
\description{
Consecutive samples representing time points in an ordination plot can be interpreted
as vectors, which have a length and an angle. If ordinate is true, these lengths and angles
are computed. If ordinate is false, dissimilarities between consecutive time points are computed.
If a perturbation object is provided and plot.type hist or box is true, the histogram or box plot of jump
lengths is plotted and the significance of the difference between perturbed and non-perturbed jump lengths is
assessed with a Wilcoxon test.
}
\examples{
\dontrun{
data(david_stoolA_otus)
data(david_stoolA_metadata)
rarefaction=rarefyFilter(david_stoolA_otus,min = 10000)
rarefiedA=rarefaction$rar
days=david_stoolA_metadata[1,rarefaction$colindices]
interpA=interpolate(rarefiedA,time.vector=days,interval=1,method="stineman")
interpA[interpA<0]=0
perturbA=perturbation(times=c(80,105), durations=c(5,9))
res=tsJumpStats(interpA, plot=TRUE, perturb=perturbA, header="in stool A")
out=powerspec(res$lengths,plot=TRUE)
}
}
