% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Time Series Plot}
\usage{
tsplot(x, time.given = FALSE, num = nrow(x), sample.points = c(),
  legend = FALSE, header = "", pch = "+", lty = 1, type = "l", ...)
}
\arguments{
\item{x}{the matrix of time series}

\item{time.given}{if true, then the column names are supposed to hold the time units}

\item{num}{the number of rows to plot (starting from the first row)}

\item{sample.points}{sample.points}

\item{legend}{add a legend}

\item{header}{string added to the plot title}

\item{pch}{plot parameter (point character)}

\item{lty}{plot parameter (line type)}

\item{type}{plot parameter (plot type, e.g. dots, lines etc.)}

\item{\\dots}{Additional arguments passed to plot()}
}
\description{
Plot the time series row-wise.
}
