% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEvol.R
\name{varEvol}
\alias{varEvol}
\title{Plot mean variance versus the number of time points.}
\usage{
varEvol(x, plot = TRUE)
}
\arguments{
\item{x}{a community matrix with rows representing taxa and columns time points}

\item{plot}{do the plot}
}
\value{
a list with variances and the intersection, slope, pval and adjR2 of the linear regression of variances against time
}
\description{
Plot the mean variance for an increasing number of time points.
}
\examples{
N=20
ts=glv(N=N,generateA(N))
out=varEvol(ts)
}
